Use File::Which to get path to mutool binary in t/use.t
This allows the path detection to work properly on Windows which appends
the contents of `$ENV{PATHEXT}` to the end of any executed command.
Simply using `-r` on the path does not work.

This also switches to using Capture::Tiny to retrieve output of the
command. This does not add a dependency because Alien::Base also uses
Capture::Tiny.

Fixes <https://github.com/project-renard/p5-Alien-MuPDF/issues/18>.